Refactor AnimalCardsForAdoption and add UserAdoptModal
Moved AnimalCardsForAdoption and its CSS to a new AnimaListComponents directory, updated imports, and added a UserAdoptModal component for user interaction when expressing interest in adoption. Updated AnimalList page to use the new component path. The new modal allows users to edit their profile information when showing interest in an animal.