Refactor user update to support partial updates
Updated the user update handler and model to allow partial updates by accepting only the fields provided in the request. The backend now dynamically builds the update object and hashes the password only if it is present. Also fixed a minor JSX issue in UserEditModal by removing an extra closing div.